functions/pdf-ingest/requirements.txt
functions-framework==3.*
google-cloud-storage
langchain-community
langchain-text-splitters
langchain-google-vertexai
langchain-postgres
asyncpg
pgvector
pypdf
google-cloud-secret-manager
